Clean up settings UI package boundaries
