+++ Packages/CmuxSettingsUI/Sources/CmuxSettingsUI/SettingsNavigationRequest.swift
@@ -1,0 +1,52 @@
+import Foundation
+
+public enum SettingsNavigationRequest {
+    public static let notificationName = Notification.Name("cmux.settings.navigate")
+    private static let targetKey = "target"
+    private static let anchorKey = "anchor"
+    private static let highlightKey = "highlight"
+
+    public static func post(_ target: SettingsNavigationTarget, anchorID: String? = nil, highlight: Bool = false) {
+        NotificationCenter.default.post(
+            name: notificationName,
+            object: nil,
+            userInfo: [
+                targetKey: target.rawValue,
+                anchorKey: anchorID ?? SettingsSearchIndex.sectionID(for: target),
+                highlightKey: highlight
+            ]
+        )
+    }
+
+    public static func target(from notification: Notification) -> SettingsNavigationTarget? {
+        destination(from: notification)?.target
+    }
+
+    public static func destination(from notification: Notification) -> SettingsNavigationDestination? {
+        guard
+            let rawValue = notification.userInfo?[targetKey] as? String,
+            let target = SettingsNavigationTarget(rawValue: rawValue)
+        else {
+            return nil
+        }
+        let anchorID = notification.userInfo?[anchorKey] as? String
+        let shouldHighlight = notification.userInfo?[highlightKey] as? Bool ?? false
+        return SettingsNavigationDestination(
+            target: target,
+            anchorID: anchorID ?? SettingsSearchIndex.sectionID(for: target),
+            shouldHighlight: shouldHighlight
+        )
+    }
+}
+
+public struct SettingsNavigationDestination: Sendable {
+    public let target: SettingsNavigationTarget
+    public let anchorID: String
+    public let shouldHighlight: Bool
+
+    public init(target: SettingsNavigationTarget, anchorID: String, shouldHighlight: Bool) {
+        self.target = target
+        self.anchorID = anchorID
+        self.shouldHighlight = shouldHighlight
+    }
+}
